+++ src/main/groovy/br/com/bluesoft/bee/model/Table.groovy
@@ -39,12 +39,13 @@ import com.fasterxml.jackson.annotation.JsonAutoDetect
 import groovy.transform.AutoClone
 
 @AutoClone
-@JsonAutoDetect(isGetterVisibility = JsonAutoDetect.Visibility.NONE, getterVisibility = JsonAutoDetect.Visibility.ANY)
+@com.fasterxml.jackson.annotation.JsonAutoDetect(isGetterVisibility = JsonAutoDetect.Visibility.NONE, getterVisibility = JsonAutoDetect.Visibility.ANY)
 class Table implements Validator {
 
     String name
     Boolean temporary = false
     String comment
+    String distStyle
 
     Map<String, TableColumn> columns = [:] as LinkedHashMap
     Map<String, Index> indexes = [:]