+++ src/main/groovy/br/com/bluesoft/bee/schema/BeeRedshiftSchemaCreatorAction.groovy
@@ -2,12 +2,13 @@ package br.com.bluesoft.bee.schema
 
 import br.com.bluesoft.bee.importer.JsonImporter
 import br.com.bluesoft.bee.model.Options
+import br.com.bluesoft.bee.model.Schema
 import br.com.bluesoft.bee.runner.ActionRunner
 import br.com.bluesoft.bee.service.BeeWriter
 import br.com.bluesoft.bee.service.RulesConverter
 import br.com.bluesoft.bee.util.RDBMS
 
-public class BeeOracleSchemaCreatorAction implements ActionRunner {
+public class BeeRedshiftSchemaCreatorAction implements ActionRunner {
 
     Options options
     BeeWriter out
@@ -24,11 +25,10 @@ public class BeeOracleSchemaCreatorAction implements ActionRunner {
     public boolean run() {
 
         def objectName = options.arguments[0]
-        def dataFolderPath = new File(options.dataDir.absolutePath, 'data')
 
         out.log('importing schema metadata from the reference files')
-        def schema = getImporter().importMetaData()
-        schema.rdbms = RDBMS.ORACLE
+        def schema = cleanupSchema(getImporter().importMetaData())
+        schema.rdbms = RDBMS.REDSHIFT
 
         if (objectName) {
             schema = schema.filter(objectName)
@@ -40,49 +40,19 @@ public class BeeOracleSchemaCreatorAction implements ActionRunner {
             file.delete()
         }
 
-        beeSchemaCreator = new BeeOracleSchemaCreator()
-
-        out.println("generating sequences...")
-        beeSchemaCreator.createSequences(file, schema)
+        beeSchemaCreator = new BeeRedshiftSchemaCreator()
 
         out.println("generating tables...")
         beeSchemaCreator.createTables(file, schema)
 
         out.println("generating core data...")
-        beeSchemaCreator.createCoreData(file, schema, options.dataDir.absolutePath)
+        beeSchemaCreator.createCoreData(file, schema, options.dataDir)
 
         out.println("generating constraints...")
-        beeSchemaCreator.createPrimaryKeys(file, schema)
-        beeSchemaCreator.createUniqueKeys(file, schema)
-        beeSchemaCreator.createForeignKeys(file, schema)
-        beeSchemaCreator.createCheckConstraint(file, schema)
-
-        out.println("generating indexes...")
-        beeSchemaCreator.createIndexes(file, schema)
-        beeSchemaCreator.createFunctionalIndexes(file, schema)
-        beeSchemaCreator.createBitmapIndexes(file, schema)
+//        beeSchemaCreator.createPrimaryKeys(file, schema)
 
         out.println("generating views...")
         beeSchemaCreator.createViews(file, schema)
-
-        out.println("generating user types...")
-        beeSchemaCreator.createUserTypes(file, schema)
-
-        out.println("generating packages...")
-        beeSchemaCreator.createPackages(file, schema)
-
-        out.println("generating procedures...")
-        beeSchemaCreator.createProcedures(file, schema)
-
-        out.println("generating materialized views...")
-        beeSchemaCreator.createMViews(file, schema)
-
-        out.println("generating materialized views indexes...")
-        beeSchemaCreator.createMViewIndexes(file, schema)
-
-        out.println("generating triggers...")
-        beeSchemaCreator.createTriggers(file, schema)
-
 
         def env = System.getenv()
         if (env['EDITOR']) {
@@ -90,6 +60,7 @@ public class BeeOracleSchemaCreatorAction implements ActionRunner {
             def cmd = [env['EDITOR'], file.path]
             new ProcessBuilder(env['EDITOR'], file.path).start()
         }
+
         return true
     }
 
@@ -99,5 +70,13 @@ public class BeeOracleSchemaCreatorAction implements ActionRunner {
         }
         return importer
     }
+
+    Schema cleanupSchema(Schema schema) {
+        schema.userTypes.clear()
+        schema.packages.clear()
+        schema.tables = schema.tables.findAll { !it.value.temporary }
+        return schema
+    }
+
 }